cast the enum to an int for cout

+++ torch/csrc/cuda/nccl.cpp
@@ -27,7 +27,7 @@ namespace detail {
 void throw_nccl_error(torchNcclResult_t status) {
   std::ostringstream err;
   ncclResult_t nccl_status = static_cast<ncclResult_t>(status);
-  err << "NCCL Error " << status << ": " << ncclGetErrorString(nccl_status);
+  err << "NCCL Error " << static_cast<int>(status) << ": " << ncclGetErrorString(nccl_status);
   throw std::runtime_error(err.str());
 }
